Move lib/shell to session/shell
