+++ session/shell/shell_test.go
