+++ session/shell/shell_unix.go
