+++ session/shell/shell_windows.go
